totsworld-main/requirements.txt
# Django==5.0.2
djangorestframework==3.15.1
pillow==10.2.0